+++ src/smtp/send.rs
@@ -3,6 +3,8 @@
 use super::Smtp;
 use async_smtp::*;
 
+use std::time::Duration;
+
 use crate::context::Context;
 use crate::events::Event;
 
@@ -30,7 +32,7 @@ impl Smtp {
         message: Vec<u8>,
         job_id: u32,
     ) -> Result<()> {
-        let message_len = message.len();
+        let message_len = message.len(); // bytes
 
         let recipients_display = recipients
             .iter()
@@ -47,7 +49,9 @@ impl Smtp {
         );
 
         if let Some(ref mut transport) = self.transport {
-            transport.send(mail).await.map_err(Error::SendError)?;
+            // The timeout is 1min + 3min per MB.
+            let timeout = 60 + (180 * message_len / 1_000_000) as u64;
+            transport.send_with_timeout(mail, Some(&Duration::from_secs(timeout))).await.map_err(Error::SendError)?;
 
             context.call_cb(Event::SmtpMessageSent(format!(
                 "Message len={} was smtp-sent to {}",